# tratamientos/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/treatments/treatment": {
            "put": {
                "description": "Updates a treatment changing everything on it, except the id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Treatment request"
                ],
                "summary": "Updates a treatment",
                "parameters": [
                    {
                        "description": "TBD",
                        "name": "treatment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Treatment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Treatment"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorMsg"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a treatment for a given animal",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Treatment request"
                ],
                "summary": "Creates a treatment",
                "parameters": [
                    {
                        "description": "TBD",
                        "name": "treatment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Treatment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Treatment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorMsg"
                        }
                    }
                }
            }
        },
        "/treatments/treatment/comment/{treatmentId}": {
            "post": {
                "description": "Adds a comment to the treatment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Treatment request"
                ],
                "summary": "Add a comment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "treatment affected",
                        "name": "pet",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Comment from the treatment",
                        "name": "comment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.CommentInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Treatment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorMsg"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorMsg"
                        }
                    }
                }
            }
        },
        "/treatments/treatment/pet/{pet}": {
            "get": {
                "description": "Updates a treatment only changing the specified fields",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Treatment request"
                ],
                "summary": "Get a treatment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "treatment affected",
                        "name": "pet",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controller.Treatment"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorMsg"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorMsg"
                        }
                    }
                }
            }
        },
        "/treatments/treatment/specific/{id}": {
            "get": {
                "description": "Updates a treatment only changing the specified fields",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Treatment request"
                ],
                "summary": "Get a treatment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "treatment affected",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Treatment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorMsg"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorMsg"
                        }
                    }
                }
            }
        },
        "/treatments/treatment/{id}": {
            "delete": {
                "description": "Removes a given treatment with given id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Treatment request"
                ],
                "summary": "Deletes a treatment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "treatment affected",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "pet affected",
                        "name": "pet",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Treatment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorMsg"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorMsg"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates a treatment only changing the specified fields",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Treatment request"
                ],
                "summary": "Updates a treatment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "treatment affected",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "TBD",
                        "name": "treatment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Treatment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Treatment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorMsg"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorMsg"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "controller.Comment": {
            "type": "object",
            "properties": {
                "date_added": {
                    "type": "string",
                    "example": "2023-01-02T15:04:05Z"
                },
                "information": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                }
            }
        },
        "controller.CommentInput": {
            "type": "object",
            "required": [
                "comment"
            ],
            "properties": {
                "comment": {
                    "type": "string"
                }
            }
        },
        "controller.ErrorMsg": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "controller.Treatment": {
            "type": "object",
            "properties": {
                "applied_to": {
                    "type": "integer",
                    "example": 20
                },
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controller.Comment"
                    }
                },
                "date_end": {
                    "type": "string",
                    "example": "2023-01-02T15:04:05Z"
                },
                "date_start": {
                    "type": "string",
                    "example": "2006-01-02T15:04:05Z"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "e013f973-ed95-45c5-8bc4-3abf2d9112c3"
                },
                "next_dose": {
                    "type": "string",
                    "example": "2023-01-02T15:04:05Z"
                },
                "type": {
                    "type": "string",
                    "example": "papota"
                }
            }
        }
    }
}